refactor(cli): add resolveMode and routed entry point
